fix(option-monitor): runtime fixes discovered during testing
- InstrumentInfo: contract_size i64→f64 (Deribit returns float 1.0)
- Connection: batch subscribe in chunks of 100 to avoid oversized WS messages
- Connection: batch reconnect re-subscribe similarly
- ConnectionPool: auto-spawn eventloop for dynamically created connections
- main.rs: spawn connection eventloops BEFORE initialize() to avoid deadlock
- main.rs: use pool.cancel_token() for unified lifecycle management

+++ crates/nq-deribit/src/model/instrument.rs
@@ -123,7 +123,7 @@ pub struct InstrumentInfo {
     pub creation_timestamp: u64,
     pub expiration_timestamp: u64,
     pub tick_size: f64,
-    pub contract_size: i64,
+    pub contract_size: f64,
     pub state: String,
     #[serde(default)]
     pub strike: Option<f64>,
